abstraher ut ferdigstillBrev component
